Refactor: Merged list and card component logic into shared components
