+++ src/components/commons/list-offers/index.ts
@@ -1,0 +1,1 @@
+export {ListOffers} from './list-offers';
