+++ src/components/commons/map/map.tsx
@@ -2,7 +2,7 @@ import { useRef, useEffect } from 'react';
 import { Icon, Marker, layerGroup } from 'leaflet';
 import useMap from '../../../hooks/use-map';
 import { City } from '../../../types/city';
-import { MAP_HEIGHT, MapMarkerSVG } from '../../../const/routes';
+import { MapMarkerSVG } from '../../../const/routes';
 import 'leaflet/dist/leaflet.css';
 import { OfferType } from '../../../types/offer-preview';
 
@@ -11,6 +11,7 @@ type MapProps = {
   city: City;
   offers: OfferType[];
   hoveredOfferId: OfferType['id'] | null;
+  className: string;
 };
 
 const defaultCustomIcon = new Icon({
@@ -26,7 +27,7 @@ const currentCustomIcon = new Icon({
 });
 
 export const Map = (props: MapProps): JSX.Element => {
-  const { city, offers, hoveredOfferId } = props;
+  const { city, offers, hoveredOfferId, className } = props;
 
   const mapRef = useRef(null);
   const map = useMap({ mapRef, city });
@@ -55,7 +56,7 @@ export const Map = (props: MapProps): JSX.Element => {
     }
   }, [map, offers, hoveredOfferId]);
 
-  return <section className="cities__map map" style={{ height: MAP_HEIGHT }} ref={mapRef}></section>;
+  return <section className={`${className} map`} ref={mapRef}></section>;
 };
 
 export default Map;
