+++ src/components/commons/offer-card/index.ts
@@ -1,0 +1,1 @@
+export {OfferCard} from './offer-card';
