--- src/components/main/list-offers/index.ts
@@ -1,1 +1,0 @@
-export { ListOffers } from './list-offers';
