--- src/components/main/offer-card/index.ts
@@ -1,1 +1,0 @@
-export { OfferCard } from './offer-card';
