--- src/components/offer/list-offers-nearby/index.ts
@@ -1,1 +1,0 @@
-export { ListOffersNearby } from './list-offers-nearby';
