--- src/components/offer/offer-card-nearby/index.ts
@@ -1,1 +1,0 @@
-export { OfferCardNearby } from './offer-card-nearby';
