--- src/components/offer/offer-map/index.ts
@@ -1,1 +1,0 @@
-export { OfferMap } from './offer-map';
